scripts/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb==0.5.20
google-generativeai==0.8.3
pydantic==2.9.2
httpx==0.27.2
